renamed a file and analyzed current behaviors before integration
